Module simple_pallet::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

advance_currency_usage

Or it can tap into the default FRAME currency, although as constrained by flite. For example, Balance is still u128, RuntimeHoldReason is still &static str adn so on.

dispatching

Warning: Doc-Only

simple_currency_usage

simple-pallet can use the very very basic flite currency primitives.